Integrate with the Feature Policy specification.
The integration with the Feature Policy specification allows us to
disallow (by default) access to this feature by cross-origin iframes
with a standardized mechanism for the top-level document to grant access
to this feature to origins that it trusts.

With this mitigation in place the Allowed Origins descriptors are
removed from the specification. This resolves the question in #49 of
whether access to USB devices should be controlled by the vendor or the
user in the favor of the user.

This resolves issue #82 and obsoletes issues #15 and #38.