Use a "requestdevice" event for flexible request handling
This change replaces the "lastFilters" and "chosenDevice" attributes
with a "requestdevice" event that is fired when requestDevice() is
called, allowing the test script to respondWith() the device it wants
the UA to simulate the user choosing.

This removes some synchronous behavior from the testing API and aligns
better with patterns used elsewhere in the web platform.