Split USB into WindowUSB and WorkerUSB
This change adds a base WindowOrWorkerUSB interface that contains all of
the existing USB interfaces minus requestDevice. Two derived interfaces
use WindowOrWorkerUSB.

WindowUSB adds the requestDevice function, and it will only be used
within the window context. WorkerUSB is the same as WindowOrWorkerUSB,
and will only be used within a worker context.