Fix up device filter matching algorithm
This change cleans up the device filter matching algorithm and reorders
the steps so that the serial number (if part of the filter) is checked
before attempting to match an interface. Otherwise the serial number
will be ignored if an interface class is part of the filter.

This resolves issue 138.